requirements.txt
line-bot-sdk==1.8.0
requests==2.19.1
gunicorn==19.8.1
codecov==2.0.15
flask==1.0.2